Added net and movement, points, check when ball is in net
Created the net - Tried to do it as a class at first but it didn't work.
needs work.
Movement - using arrows and identifying walls of the arena.
Check when ball is in net - Wasn't able to take it out of the arena
though, so colored it in black in the meantime.
Points - Created a property in Ball class which contains points amount,
and in Main.js the points raise when a ball enters the net.